Move ClientLayout from app to layouts
