+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import { Inter, Montserrat } from 'next/font/google';
 import './globals.css';
-import ClientLayout from './ClientLayout';
+import ClientLayout from '../components/layouts/ClientLayout';
 
 const inter = Inter({
   subsets: ['latin'],
